orchestrator/email_coordinator/draft_manager/student_mapper/requirements.txt
openpyxl>=3.1.0
pyyaml>=6.0
pytest>=7.0.0
